refactor(map): extract layer config, hover UI, keyboard nav, and traceroute plumbing from Map.tsx into focused modules
